[feature/JMT-103] feat : S3 이미지 삭제 부분에 default 이미지는 삭제하지 않도록 변경

+++ src/main/java/com/gdsc/jmt/global/service/S3FileService.java
@@ -31,15 +31,23 @@ public class S3FileService {
     }
     public void delete(String path) {
         DeleteObjectRequest request = getDeleteObjectRequest(path);
-        s3Client.deleteObject(request);
+        if(request == null) {
+            return;
+        }
+         s3Client.deleteObject(request);
     }
 
     private DeleteObjectRequest getDeleteObjectRequest(String path) {
+        String key = path.substring(path.lastIndexOf("com/") + 4);
+        if(key.equals("profileImg/defaultImg/Default+image.png")) {
+            return null;
+        }
         return DeleteObjectRequest.builder()
                 .bucket(bucket)
-                .key(path.substring(path.lastIndexOf("com/") + 4))
+                .key(key)
                 .build();
     }
+
     public String upload(MultipartFile multipartFile, String path) throws IOException {
         String originName = multipartFile.getOriginalFilename();
         //확장자 추출
